refactor: split server into modular routers
